+++ R/get_dataset_list.R
@@ -11,6 +11,11 @@ get_dataset_list = function(){
   link_dataflow = Sys.getenv("INSEE_sdmx_link_dataflow")
 
   df = get_insee(link_dataflow)
+  if(!is.null(df)){
+    df = dplyr::filter(.data = df, !is.na(.data$Name.fr))
+  }
+
+  # df = tibble::as_tibble(df)
 
   return(df)
 }