+++ R/get_insee.R
@@ -5,11 +5,11 @@
 #' It is used by the functions get_insee_dataset, get_insee_idbank and get_dataset_list.
 #' The data is cached, hence all queries are only run once per R session.
 #' The user can disable the download display in the console with the following command :
-#' Sys.setenv(INSEE_download_verbose = "FALSE")
-#' The use of cached data can be disabled with : Sys.setenv(INSEE_no_cache_use = "TRUE")
+#' Sys.setenv(INSEE_download_verbose = "FALSE").
+#' The use of cached data can be disabled with : Sys.setenv(INSEE_no_cache_use = "TRUE").
 #' All queries are printed in the console with this command: Sys.setenv(INSEE_print_query = "TRUE").
-#' By default the RapidXML C++ library is used through the readsdmx package.
-#' The previous parser can still be used with this command : Sys.setenv(INSEE_read_sdmx_slow = "TRUE")
+#' The RapidXML C++ library is used as a backup thanks to the readsdmx package.
+#' It can be used instead of the internal parser with this command : Sys.setenv(INSEE_read_sdmx_fast = "TRUE")
 #' @param link SDMX query link
 #' @param step argument used only for internal package purposes to tweak download display
 #' @return a tibble containing the data
@@ -31,6 +31,7 @@ get_insee = function(link, step = "1/1"){
   insee_value_as_numeric = if(Sys.getenv("INSEE_value_as_numeric") == "TRUE"){TRUE}else{FALSE}
   insee_print_query = if(Sys.getenv("INSEE_print_query") == "TRUE"){TRUE}else{FALSE}
   insee_no_cache_use = if(Sys.getenv("INSEE_no_cache_use") == "TRUE"){TRUE}else{FALSE}
+  insee_read_sdmx_fast = if(Sys.getenv("INSEE_read_sdmx_fast") == "TRUE"){TRUE}else{FALSE}
 
   if(insee_download_verbose){
     if(insee_print_query == TRUE) {
@@ -43,33 +44,38 @@ get_insee = function(link, step = "1/1"){
 
   if((!file.exists(file_cache)) | insee_no_cache_use){
 
-    use_read_sdmx_fast = TRUE
-
-    if(Sys.getenv("INSEE_read_sdmx_slow") == "TRUE"){
-      use_read_sdmx_fast = FALSE
-    }
+    use_read_sdmx_fast_first = insee_read_sdmx_fast
+    use_backup_parser = TRUE
 
     if(stringr::str_detect(link, "includeHistory")){
-      use_read_sdmx_fast = FALSE
+      use_backup_parser = FALSE
+      insee_read_sdmx_fast = FALSE
     }
 
     if(link == Sys.getenv("INSEE_sdmx_link_dataflow")){
-      use_read_sdmx_fast = FALSE
+      use_backup_parser = FALSE
+      insee_read_sdmx_fast = FALSE
     }
 
-    if(use_read_sdmx_fast){
-      data_final = read_sdmx_fast(link, step)
-    }
-
-    # read_sdmx_slow is used as a backup solution in case read_sdmx_fast is not working
-    if(use_read_sdmx_fast == TRUE){
-      if(is.null(data_final)){
-        use_read_sdmx_fast = FALSE
-      }
-    }
-
-    if(use_read_sdmx_fast == FALSE){
+    # by default the internal parser is used, if it fails the readsdmx parser is used
+    if(!use_read_sdmx_fast_first){
       data_final = read_sdmx_slow(link, step)
+
+      if(use_backup_parser){
+        if(is.null(data_final)){
+          data_final = read_sdmx_fast(link, step)
+        }
+      }
+
+    }else{
+      data_final = read_sdmx_fast(link, step)
+
+      if(use_backup_parser){
+        if(is.null(data_final)){
+          data_final = read_sdmx_slow(link, step)
+        }
+      }
+
     }
 
     if(!is.null(data_final)){
@@ -77,14 +83,13 @@ get_insee = function(link, step = "1/1"){
       saveRDS(data_final, file = file_cache)
 
       if(insee_download_verbose){
-        if(use_read_sdmx_fast){
-          msg = sprintf("Data cached : %s\n", file_cache)
-        }else{
-          msg = sprintf("\nData cached : %s\n", file_cache)
-        }
+        msg = sprintf("Data cached : %s\n", file_cache)
 
         message(crayon::style(msg, "green"))
       }
+    }else{
+      msg = "An error occurred"
+      message(crayon::style(msg, "red"))
     }
   }else{